Add Favorites table

+++ app/controllers/user_manager.py
@@ -1,6 +1,7 @@
 import json
 from datetime import datetime, timedelta
 from flask import jsonify
+from flask import request
 
 from app import db
 from app.config.constants import ErrorCodes
@@ -76,7 +77,7 @@ def user_login(data) -> json:
             session_old.uuid = temp_uuid
             session_old.update_session()
         else:
-            session = Session(user.username, temp_uuid, datetime.utcnow())
+            session = Session(username, temp_uuid, datetime.utcnow(), request.remote_addr)
             session.save()
         return generate_login_success_response(temp_uuid)
     else:
@@ -85,14 +86,17 @@ def user_login(data) -> json:
 
 def is_user_login(uuid: str) -> str:
     now = datetime.utcnow()
+    remote_address = request.remote_addr
     session = db.session.query(Session).filter_by(uuid=uuid).first()
-    updated_time = session.time_stamp + timedelta(minutes=30)
-    if updated_time < now:
-        session.delete()
-    else:
-        session.time_stamp = now
-        session.update_session()
-        return session.username
+    if remote_address == session.remote_address:
+        if session:
+            updated_time = session.time_stamp + timedelta(minutes=30)
+            if updated_time < now:
+                session.delete()
+            else:
+                session.time_stamp = now
+                session.update_session()
+                return session.username
 
 
 @validate_schema(the_admin_login_schema)
@@ -113,7 +117,7 @@ def the_admin_login(data):
             session_old.uuid = temp_uuid
             session_old.update_session()
         else:
-            session = Session(username, temp_uuid, datetime.utcnow())
+            session = Session(username, temp_uuid, datetime.utcnow(), request.remote_addr)
             session.save()
         return generate_login_success_response(temp_uuid)
     else:
